refactor(webview): remove duplicated lifecycle methods, move to lifecycle.py (Iteration #131)
