+++ python/auroraview/core/mixins/__init__.py
@@ -21,6 +21,7 @@ from auroraview.core.mixins.events import WebViewEventMixin
 from auroraview.core.mixins.api import WebViewApiMixin
 from auroraview.core.mixins.dom import WebViewDOMMixin
 from auroraview.core.mixins.telemetry import WebViewTelemetryMixin
+from auroraview.core.mixins.lifecycle import WebViewLifecycleMixin
 
 __all__ = [
     "WebViewWindowMixin",
@@ -30,4 +31,5 @@ __all__ = [
     "WebViewApiMixin",
     "WebViewDOMMixin",
     "WebViewTelemetryMixin",
+    "WebViewLifecycleMixin",
 ]
